requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic
python-dotenv
python-multipart
requests
# 👇 NEW AI SDKs (Critical for Phase 6)
google-generativeai
openai
anthropic